requirements.txt
fastapi
uvicorn
langchain
langchain-huggingface
faiss-cpu
pypdf
python-dotenv
sentence-transformers
streamlit
pydantic
requests
